requirements.txt
# Spring 2025 Authors: Braden Doty
Flask
Flask-WTF
email_validator
pandas
pycryptodome
gunicorn
APScheduler
